+++ src/aws/pci-dss-3.2.1/rules/index.ts
@@ -1,6 +1,6 @@
 import Aws_PCI_DSS_321_Autoscaling_1 from './pci-dss-3.2.1-autoscaling-check-1'
-import Aws_PCI_DSS_321_Alb_1 from './pci-dss-3.2.1-alb-check-1'
 import Aws_PCI_DSS_321_Cloudfront_1 from './pci-dss-3.2.1-cloudfront-check-1'
+import Aws_PCI_DSS_321_Elbv2_1 from './pci-dss-3.2.1-elbv2-check-1'
 import Aws_PCI_DSS_321_Codebuild_1 from './pci-dss-3.2.1-codebuild-check-1'
 import Aws_PCI_DSS_321_Codebuild_2 from './pci-dss-3.2.1-codebuild-check-2'
 import Aws_PCI_DSS_321_Cloudtrail_1 from './pci-dss-3.2.1-cloudtrail-check-1'
@@ -11,6 +11,11 @@ import Aws_PCI_DSS_321_Dms_1 from './pci-dss-3.2.1-dms-check-1'
 import Aws_PCI_DSS_321_IAM_1 from './pci-dss-3.2.1-iam-check-1'
 import Aws_PCI_DSS_321_IAM_2 from './pci-dss-3.2.1-iam-check-2'
 import Aws_PCI_DSS_321_IAM_3 from './pci-dss-3.2.1-iam-check-3'
+import Aws_PCI_DSS_321_IAM_4 from './pci-dss-3.2.1-iam-check-4'
+import Aws_PCI_DSS_321_IAM_5 from './pci-dss-3.2.1-iam-check-5'
+import Aws_PCI_DSS_321_IAM_6 from './pci-dss-3.2.1-iam-check-6'
+import Aws_PCI_DSS_321_IAM_7 from './pci-dss-3.2.1-iam-check-7'
+import Aws_PCI_DSS_321_IAM_8 from './pci-dss-3.2.1-iam-check-8'
 import Aws_PCI_DSS_321_EC2_1 from './pci-dss-3.2.1-ec2-check-1'
 import Aws_PCI_DSS_321_EC2_2 from './pci-dss-3.2.1-ec2-check-2'
 import Aws_PCI_DSS_321_EC2_4 from './pci-dss-3.2.1-ec2-check-4'
@@ -24,7 +29,6 @@ import Aws_PCI_DSS_321_RDS_2 from './pci-dss-3.2.1-rds-check-2'
 
 export default [
   Aws_PCI_DSS_321_Autoscaling_1,
-  Aws_PCI_DSS_321_Alb_1,
   Aws_PCI_DSS_321_Cloudfront_1,
   Aws_PCI_DSS_321_Codebuild_1,
   Aws_PCI_DSS_321_Codebuild_2,
@@ -36,11 +40,17 @@ export default [
   Aws_PCI_DSS_321_IAM_1,
   Aws_PCI_DSS_321_IAM_2,
   Aws_PCI_DSS_321_IAM_3,
+  Aws_PCI_DSS_321_IAM_4,
+  Aws_PCI_DSS_321_IAM_5,
+  Aws_PCI_DSS_321_IAM_6,
+  Aws_PCI_DSS_321_IAM_7,
+  Aws_PCI_DSS_321_IAM_8,
   Aws_PCI_DSS_321_EC2_1,
   Aws_PCI_DSS_321_EC2_2,
   Aws_PCI_DSS_321_EC2_4,
   Aws_PCI_DSS_321_EC2_5,
   Aws_PCI_DSS_321_EC2_6,
+  Aws_PCI_DSS_321_Elbv2_1,
   Aws_PCI_DSS_321_SageMaker_1,
   Aws_PCI_DSS_321_ES_1,
   Aws_PCI_DSS_321_ES_2,